refactor(HttpMessageCodec): Split in multiple files (#131)
