+++ zio-http/src/main/scala/zhttp/service/DecodeJRequest.scala
@@ -1,0 +1,18 @@
+package zhttp.service
+
+import zhttp.core.JFullHttpRequest
+import zhttp.http._
+
+trait DecodeJRequest {
+
+  /**
+   * Tries to decode the [io.netty.handler.codec.http.FullHttpRequest] to [Request].
+   */
+  def decodeJRequest(jReq: JFullHttpRequest): Either[HttpError, Request] = for {
+    url <- URL.fromString(jReq.uri())
+    method   = Method.fromJHttpMethod(jReq.method())
+    headers  = Header.make(jReq.headers())
+    endpoint = method -> url
+    data     = Request.Data(headers, HttpContent.Complete(jReq.content().toString(HTTP_CHARSET)))
+  } yield Request(endpoint, data)
+}
